Add properties and normalization

+++ NTDataHiveGrpcService/DAL/Model/Revision.cs
@@ -9,8 +9,8 @@
         public string PublisherName { get; set; }
         public string JournalId { get; set; }
         public string ArticleId { get; set; }
-        public int PageCount { get; set; }
-        public int ErrorCount { get; set; }
+        public double PageCount { get; set; }
+        public double ErrorCount { get; set; }
         public string DescriptionOfError { get; set; }
         public string Matter { get; set; }
         public string ErrorLocation { get; set; }
